requirements.txt
Flask==3.0.3
itsdangerous==2.2.0
Pillow==10.4.0
